requirements.txt
google-generativeai
pymongo
SpeechRecognition
PyAudio  # Cần thiết cho SpeechRecognition để truy cập microphone
dotenv